C#: Convert times and vectors to/from strings.
Implemented and added tests for the following methods:

    AstroTime.TryParse()
    AstroVector.TryParse()

I still need to implement StateVector.TryParse().

Fixed a bug in CalendarDateTime: the NOVAS cal_date
function worked well, except when years go below
somewhere near -12000. Then the formulas start making
negative numbers, which messes up the calculation
of the month and day.

So to fix this, I figured out that any multiple of
400 years added to any calendar date gives the exact
same calendar date. And 400 years always has the exact
same number of days in it: 146097.

Because one million years = 400 * 2500 years, I can
add 2500*146097 days at the front of the formula
and subtract 1000000 years at the back, and everything
works for the entire range of years plus or minus
one million years from the present.

Because my date format allows for no more than a 6-digit
decimal year, this is perfectly adequate.