I had a build failure in Windows due to LF/CRLF in Go source.
I'm not sure this is going to fix the error, because I'm not
sure how 'go fmt' works. I'm hoping that if I fix the newlines
in the generated code, then format it, under Windows it will
end up exactly like what git does to normalize the line endings
when we clone the repo.