Attempting to diagnose recent changes in Python calculations.
It looks like values returned by VsopModel() are inconsistent
running the same code on different hardware.
On my system, Python and C are producing nearly identical results,
but in GitHub Actions, I'm getting errors on the order of 1e-12.

I'm beginning a series of commits in a separate branch where
I can hack the code to debug my code where it fails, which is
unfortunately only on GitHub Actions.